Polished some things, added some others tweaked the gps stuff a little changed server urls to hit the production server added a submit activity, and associated view only accessible from the map activity tweaked the op list item view a little changed the reply adapter... we no longer need the view generators to be in seperate static classes, but they are still there anyway the reply list view now has a big space for the OP message

+++ app/src/main/java/groovycarnage/com/hermes/adapters/AdapterViewGenerators.java
@@ -40,8 +40,8 @@ public class AdapterViewGenerators {
 
         content.setText(message.content);
         poster.setText("username here");
-        latitude.setText(Float.toString(message.lat));
-        longitude.setText(Float.toString(message.lon));
+        latitude.setText(Double.toString(message.lat));
+        longitude.setText(Double.toString(message.lon));
 
         return itemView;
     }
